requirements.txt
numpy==1.24.3
pandas==2.0.3
Pillow==9.4.0
streamlit==1.29.0
